+++ Sanctuary.Census/Models/Collections/LoadoutSlot.cs
@@ -1,0 +1,28 @@
+﻿using Sanctuary.Census.Attributes;
+using Sanctuary.Census.Common.Objects.CommonModels;
+using System.ComponentModel.DataAnnotations;
+
+namespace Sanctuary.Census.Models.Collections;
+
+/// <summary>
+/// Represents a slot of an infantry loadout.
+/// </summary>
+/// <param name="LoadoutID">The ID of the loadout that this slot belongs to.</param>
+/// <param name="SlotID">The ID of the slot.</param>
+/// <param name="Name">The locale string ID of the slot's name.</param>
+/// <param name="Description">The locale string ID of the slot's description.</param>
+/// <param name="IsAutoEquipped">Whether this slot is automatically equipped.</param>
+/// <param name="IsRequired">Whether this slot is required to be equipped.</param>
+/// <param name="UiTag">The UI section that this slot appears under.</param>
+[Collection]
+public record LoadoutSlot
+(
+    [property: Key] uint LoadoutID,
+    [property: Key] uint SlotID,
+    LocaleString? Name,
+    LocaleString? Description,
+    bool IsAutoEquipped,
+    bool IsRequired,
+    bool IsVisible,
+    string? UiTag
+);